Read the completion counts at their full width in the final-road test
Widening ds:0x452a to uint16_t left the final-road flag iterating the array
as uint8_t, which truncates: a count of 256 would have read as zero and
wrongly made a finished road look untouched.

+++ src/core/app.cpp
@@ -573,7 +573,7 @@ void AttractModeApp::start_gameplay(std::vector<AudioCommand>& audio) {
     {
         const std::size_t flat = go_menu_flat_index(selected_world_, selected_level_);
         std::size_t completed = 0;
-        for (uint8_t c : road_completions_) {
+        for (uint16_t c : road_completions_) {
             if (c != 0) completed += 1;
         }
         current_road_is_final_ =
